iopy: fix in-place dict pop on class object creation.
When using the same dict twice with "_class" to initiate two class
objects, it failed on the second creation:

        class_b_dict = {
            '_class': 'test.ClassB',
            'field1': 28,
            'field2': 78
        }

        union_a_1 = self.r.test.UnionA({
            'a': class_b_dict
        })
        union_a_2 = self.r.test.UnionA({
            'a': class_b_dict
        })

:        union_a_2 = self.r.test.UnionA({
:      File "iopy.pyx", line 1697, in iopy._InternalStructUnionMetaclass.__call__
:      File "iopy.pyx", line 5178, in iopy.UnionBase.__init__
:      File "iopy.pyx", line 5449, in iopy.union_set
:      File "iopy.pyx", line 4100, in iopy.convert_field_object
:      File "iopy.pyx", line 4014, in iopy.explicit_convert_field
:      File "iopy.pyx", line 1697, in iopy._InternalStructUnionMetaclass.__call__
:      File "iopy.pyx", line 5555, in iopy.StructBase.__init__
:    iopy.Error: Invalid key field2, allowed: field1, optField

This was because on class object creation, the given dict was modified
in-placed and the '_class' keyword was removed.

The solution is to do a copy of the dict on object creation in case of
'_class' keyword before removing it.

Change-Id: Ibbe8664ae9b19b7108f51ce7d3f93f0cff8875ed

rip-it: 5b3fb86fadb5f29915075df9b748deeacf9c3fe7