waf/backend: improve llvm/clang validation.
Lib-common only supports llvm version >= 9 and <= 11 for now.
Use latest supported version available if possible.
Verify and raise an error if the llvm version is not supported.

Verify and raise an error if clang or clang++ are not found.

Change-Id: Ib5d0a20350c07e849699c8f1724bf3b69f1df6c8

rip-it: f2dd6e7d8a41464ad6379f05a109e094e405f780