pxcc: fix using `raise` in pxcc tests that conflicts with raise(3).
In 9cee470003943132b we check that different Python keywords are well
escaped.
Unfortunately, we used `raise` to be defined as a union.
`raise` is already a symbol defined by raise(3).
So an error was raised when <signal.h> was included in the same C file
which is the case for Python3.5.

The solution is to avoid using `raise` as a symbol. Use `elif` instead.

Change-Id: I5f6a5156c48aef3da48632ff65ec95085d643ed8

rip-it: 54cd411b1a5409a63955144c0c4cc8d78cff6f03