waf: fix race condition with depends_on and custom_install.
When a custom task gen was depending on another task gen with
`depends_on` keyword, we would expect the `custom_install` commands of
the first task gen to be run after the `custom_install` commands of the
second task gen.

But it was not actually the case and the two commands were run without
any ordering leading to some race issues.

The issue was that when we created the tasks for the `custom_install`
commands, we did not actually consider the `depends_on` relations.

The solution is when we create the `custom_install` tasks, consider all
the tasks of the `depends_on` task gens as dependencies.
This way we have a nice ordering of `custom_install` commands between
all the tasks gen.

Change-Id: I7a440a5613a3b3b7157b7139027545bcffe0dada

rip-it: 19679f453d6fde56836f84c172b2c3b778fd8571