Synchronized on private 2019.3 branch.
Due to some commits and backports in older non-public branches, it is
not possible to automatically synchronize the public branch.

The solution is to do it manually.

Here is the list of commits included by this synchronization:

Nicolas Pauss (6):
      396f0b527 [BACKPORT] cflags: fix compatibility with Clang 11.
      83a412f23 lib-common: fix build on clang 12.
      99334d284 parseopt: fix global-buffer-overflow on copyinit().
      4c4670dbe parseopt: fix using get_int_init() for OPTION_FLAG.
      e8e5cb60f [BACKPORT] zchk/parseopt: add tests for parseopt function.
      7e5ffbbaa zchk/parseopt: check '--no-' args and initial copy of correct int size.

Romain Le Godais (1):
      2ae57c4f2 [BACKPORT] core-mem: Fix possible link error on __imalloc_too_large.

rip-it: 092a196e10999e9944595bdf8db0a964241a4516