core/errors: make sure we write the whole backtrace in .debug files
When writing .debug files, use a 'fsync()' to secure the writing
progress and make sure we print the whole backtrace even if a problem
occurs while writing the other parts of the .debug file.

Not 100% sure but it might fix a problem of truncated .debug file seen
while working on issue #82253.

Change-Id: I91f78dc49825c835d6858d325fa3d86a516118da
Priv-Id: d8ae0e4a1bb3152b9bc910b77f1981dfdc7be5f0