net/hpack: add HPACK header tables
Added code to enable HPACK table-based header compression as defined in
RFC 7541. The code includes the static table that contains the
predefined headers and method to create and populate the dynamic table
which is supposed to contain additional application-specific headers.
The code also contains fast lookup and indexing procedures needed in
the encoding & decoding procedures in the next patch.

Change-Id: If5ea41d6cb861466a8320695960a8deede53d4f4
Priv-Id: c75489f9925de798403f475f173e035f4a7b3211