waf[common]: Check circular dependencies caused by "depends_on".
Circular dependencies (between task generators) could be introduced with
the "depends_on" feature without being detected. In case of a circular
dependency, the relative order or the posts between task generators is
not deterministic anymore (it can depend on the directory from which waf
is launched for example), and it can cause some undefined behavior.

So check and forbid circular dependencies introduced by depends_on when
processing them.

Change-Id: I0c8cddfafc9156754a6da22c98c82623fdc494c5
Priv-Id: 394d1de95845633fdf9fbd9f15c5c3da233ebad2