waf: check that elements in 'use' list do exist
Before this patch, having broken elements in the 'use' list of task
generators had no consequence. This patch make sure that each element in
the list matches an existing task generator or a 'XXX_<name>' variable
from building environment.

Purpose: avoid letting typos be a threat for the time and mental health
of people that make them.

Also remove 'compat' from 'use' list of task generator
'libcommon-minimal' as it isn't related to anything.

Change-Id: I53b155cc8fea3efd100a36f70c85dee36721406d
Priv-Id: b78c69332cfea006e26d22f5dd33641974fb6e65