iopc/lang-c: unduplicate attribute dumping for enum values
This patch splits 'iopc_dump_attrs()' in two and use new function
'iopc_dump_attr_table()' for the C dumping of enum values attributes.  A
large portion of the dumping code was duplicated between functions
'iopc_enum_dump_src()' and 'iop_dump_attrs()' (former DUMP_ATTRS macro).

The code of 'iopc_dump_attrs()' had to be reworked a bit because the
'attrs' structure of all enum values of a given enum are put in an array
instead of being dumped in a dedicated global variable. The new function
'iopc_dump_attr_table()' allows to dump the 'attrs' structure separately
if needed.

Change-Id: I93af3b582803ea5be86a26f37d836fecdaaf2e5e
Priv-Id: 430ca5599a69f3d33b12129f8a1085eabefb8109