iopc/lang-c: replace all calls to 'iopc_dump_attrs_only_help()'
Use functions 'iopc_dump_attr_table()' and 'iopc_dump_attrs()' to remove
and replace each call to 'iopc_dump_attrs_only_help()': it didn't do
anything different, it just did simpler things.

The purpose is still to unify the way we write attributes in the code,
but there are still a few attribute kinds to migrate.

Change-Id: Ia8447ee90bf5f113c50021b96a6c830d2fd8b535
Refs: #79937
Priv-Id: 11ec15e9613d1ed5d72b67b989ab808990758ab0