iopy: fix deprecated special custom methods __*__.
0a10a0189100c9e has deprecated the special custom methods __*__.
To do that, we used aliases `__from_file__ = from_file`.
Unfortunately, this does not work with classmethods.

The solution is to simply keep the old deprecated methods with the same
behaviour as the new methods.

We now check the behaviour and result of the methods in tests instead of
just the aliases.

Change-Id: I58bb49bc04db1c0a41c09a17eda79e181034ec74
Priv-Id: 417b0cf4b5caf4b7abd741aefc036ff927284d85