qps: fix thread-safety issues
QPS has the capability of mapping memory just-in-time by catching the
segmentation fault signal.

The issue is that it doesn't have context to tell in which QPS the
concerned memory address was when the segmentation fault happened, so it
scans all the "allowed" maps of all existing QPS to find it, which is
absolutely not thread-safe.

This patch adds a spinlock to protect QPS against data races around each
read or modification of QPS allowed maps.

Change-Id: I1c1524d3f8b6b3c6bfc57b011c0f1ec8cc955a56
Priv-Id: dd2a378b4e2822ef1ba0190039db398f870ddbaf