cython: avoid alloca for sb_t.
Since 00513c0458, we don't use alloca to initialize a string buffer, but
we initialize it with a static buffer with a constant size.
Unfortunately, with cython we cannot hide the creation of the static
buffer.

The solution is to explicitly create the static buffer before
initialiazing the string buffer.
Thanks to some typedefs and macros, we can make sure that the only
accepted buffer are the buffer of constant sizes.

Change-Id: I80c4840228c7446703ca589009169311171ffbfd
Refs: #80373
Priv-Id: d5f76a04c71da4512aa799bd826f45d52ab8d8a8