qps-hat: fixed value pointer update in enumeration_next
qhat_tree_enumeration_next works by updating the pos in the
qhat_tree_enumator_t, finding the node and then update the value
pointer by computing the difference between old pos and new pos.
If the node was changed, old pos is not supposed to be taken in
account.
But there is an issue in the way the function check if the node has
changed.

The function works this way:
1. keep a reference to the current node (old_node)
2. ensure that enumerator is sync with the qhat
3. increment pos in the enumerator (and keep old value)
4. find node
5. check for node change
6. update value pointer

Step 6 assume that the new pos will always be greater than the
previous one. This is supposed to be garanted by step 5 which will
reset old pos if it detect that the node did change.

The issue is that while pos is linked to the node, the previous value
for pos and node are not saved at the same time and so can be
desynchronized if step 2 did change the node. When this happen we can
end up with an old position being greater than the new position.

So the fix is simple, as we don't need old_node before step 5, we can
just save it at the same time than we save the old pos at step 3.

Change-Id: Icbe94fc7f1ae1ad63c3ff4fce717df3d09d3d433
Priv-Id: ab9b6b41fbedc1cde673abd062e0688526e1ec1c