qps/qhat: add some doc into "qhat_tree_enumerator_t"
I spent some time today to try to understand what was wrong in
"qhat_tree_enumerator_t". To make it short, the enumerator is really bad
when it comes to handle cases where the qhat is modified without
changing the qhat generation, and the attribute "count" really makes it
worse.

This patch doesn't fix anything but adds some doc to explain what is
in some of the enumerator attributes, as it's not so obvious and can
spare some time to anyone who would be trying to fix stuff in qhat tree
enumerator.

Change-Id: I85148a367c53d75a328e246a4cd49a0ea5b0a506
Priv-Id: 5b6c3871b7d3cac86a29026354a1ee365e8c7a50