yaml: fix parsing next field for sequence with field with same ident.
Using the same level of indentation for sequence elements as the object
key is valid in YAML:
  a:
  - 3
  - ~
  b:4

fcec3c140ccb999e9 was done to support this case.
Unfortunately, the next field was not well handled after the sequence,
an error was raised:

 <string>:4:1: wrong type of data, expected another element of sequence
 b: 4

The solution is, instead of returning an error in that case, consider
the sequence over and parse the data as a next field.

Change-Id: I347dc2fe96952604ebf2e76e253d7342728213c2
Priv-Id: 7ce91cf455079ab494c687547e16d60d8c9f75e8