rpc-el: fix segfault when disconnecting the client on connection.
Issue
=====

Before, ic_el_client_disconnect() called ic_disconnect() directly.

This was a big issue and caused a segfault when
ic_el_client_disconnect() was called when the client was connecting and
the event loop thread was in ic_el_client_on_event().
It is not possible to call ic_disconnect() while the ichannel is in the
on_event callback.

This can easily happen now with 5232e15c4d5167240a83 that adds the
on_connect callback.
This can also happen when another thread calls ic_el_client_disconnect()
when the mutex is unlocked in ic_el_client_on_event(), for example
through ic_el_client_notify_async_connections().

This never happened before 5232e15c4d5167240a83 and 3f0dcdd716a70eb,
because we did not have a callback function and we never released the
internal mutex in ic_el_client_on_event().
This had the effect of blocking any call of ic_el_client_disconnect()
after the processing of ic_el_client_on_event().

Solution
========

The solution is to not call directly ic_disconnect() in
ic_el_client_disconnect() when the client is connecting
(client->in_connect).
With a new dlist _G.disconnecting_clients, we now disconnect the clients
with ic_disconnect() on the next iteration of the event loop and not in
ic_el_client_on_event().

Also optimize ic_el_client_notify_async_connections() to not release the
mutex or do unnecessary operations when we don't have any asynchronous
connection contexts.

Change-Id: I730ade4558d6205f9cd71a513965c4ad08bd2f7b
Refs: #63874
Priv-Id: 1fb7664df38c5fb4ec8ca23ab2b09b8dcac78ffa