iop/field-path: allow wildcard indexing for each usage of field path.
We now allow wildcard indexing in every functions that use field paths.

This permits to filter or sort on fields where we know there will only
be one single instance of a sub-class in an array of object.
For example, `e[*].<tstiop.MyClass3>int3` where we know that the array
`e` contains multiple objects but only one object of type
`tstiop.MyClass3`.

There are still some limitations:
    1. It is not possible to use wildcard indexing with dynamic class
       cast. An error is returned in that case when building the field
       path.
    2. Using wildcard indexing with a path that can return multiple
       values is undefined behaviour when filtering or sorting objects.

Change-Id: If5d8c0b5d1fe1e08916a5b19fab675f8ab3d8bdf
Refs: #83295
Priv-Id: cbf1ff6c5a8b0aecdd1c7a9b1cb3621e3e9b7b50