iop[json]: make unpacker more tolerant with USE_C_CASE + IGNORE_UNKNOWN
Previously when using USE_C_CASE + IGNORE_UNKNOWN flags, if we found an
unknown field and c_to_camelcase failed, we returned an error and the
whole json was considered invalid.

But if IGNORE_UNKNOWN is set, we can simply ignore the field which
cannot be represented in IOP anyway (for instance a field containing
'-').
If IGNORE_UNKNOWN is not set, we keep the same error as previously
(iop_field_find_by_name will fail).

Change-Id: I256efa52941bddefbcaba4bb79a94656f31564cd
Priv-Id: d6d03afeebb228886b18f96818f30f4cc13aff62