net/addr: rework addr_resolve() and add addr_resolve_with_err().
In case of error, addr_resolve() do an e_error() that is sometimes not
really wanted, it can be easier to process the error description
manually.

Add addr_resolve_with_err() to be able to get the error description and
not do an e_error().

addr_resolve() is kept as is for backward compatibility.

Use addr_resolve_with_err() in __ic_connect() instead of duplicating the
warning.

Also remove host and port from addr_resolve().
Since c2a356993f90c56dfb, out_host and out_port are nullable for
addr_resolve2(), and in addr_resolve() they are just set with temporary
variables and not actually used. So we can safely remove these variables
and use NULL instead.

Change-Id: I4b392942bcf328a87ea1b96769175747dfea9bf1
Refs: #84006
Priv-Id: d74d006b932e9831316da3f3e72350ab55aef315