lib-common/iprintf: fix buffer underflow with quoted modifier.
The size used to format the integers was precisely defined to
hold the formatting of the maximum integer value:
(64 + 2) / 3 + 1 + 1 == 24

Unfortunately, with the quoted modifier, this is broken.
With a UINT64_MAX and quoted modifier, we have 26 characters.
So, instead of being clever by finding out the precise buffer size, use
64 as the buffer size to be sure.

Change-Id: I2c5fbb99cef9ecbf735979785831cc5e6b5c9f37
Closes: #84391
Priv-Id: 6dd39e4815908a935876969967ecb35f9c6e93e7