mypy/iopy: add stub file for IOPy.
IOPy is a compiled library. So we need to add a stub file for it.

This is still not perfect.
In IOPy, we create classes and set attributes dynamically when loading a
DSO. This does not work well with mypy.

One solution would be to either generate the stubs for the DSOs or use a
mypy plugin to automatically use the correct types.

Deactivate mypy checks for z_iopy for now.

Change-Id: I750a156a517cfeb69dd2c87090f188d3f9697823
Refs: #84968
Priv-Id: af2b4ba3245169394cda93b34ccece5e1fcbd084