mypy/zpycore: improve and add full typing for zpycore.
zpycore is our library of tests. We want it to be fully typed.
Be strict on mypy errors in zpycore and add typing for all functions.

Importing all the symbols of unittest in zunittest and dynamically
setting `__all__` was not well supported by mypy.
Since mypy does not run the modules, it requires __all__ to be
statically defined.
Furthermore, some symbols of unittest are redefined by zunittest, and
pylint and mypy do not like theses conflicts with wilcard imports in
zunittest.

Fortunately, we never actually use zunittest directly, it is always used
through zpycore.
So what we can do is import all the symbols of unittest and zunittest in
zpycore/__init__.py, and the symbols of zunittest override the symbols
of unittest as wanted, and both pylint and mypy are happy.

Drop the support unittest2 which was only used by Python 2.x, which we
drop the support a long time ago.

Also fix `debug` argument support for ZTestSuite.run() (I don't think we
actually use it anyway)

Change-Id: I464113097c421bd4f70151cddd3f54b481718b84
Refs: #84968
Priv-Id: bac0eafb50977fd890e822685dd4abe93e1bb8fe