[BACKPORT] lib-common/el: fix tolerance on timer without TIMER_LOWRES flag.
The timers that have the TIMER_LOWRES flag are the only ones that should
have a non null tolerance. Also, the comments are grouped for more
clarity, and the XXX is removed since it is not relevant.

Change-Id: Ia6ce2b95f084dbbe13584af2ea7cf0e4ea68d5df
(cherry picked from commit cef56b378091b3dc987f73970e5ad2ca058f9578)
Priv-Id: 65335da78a88543192cf211c25463a3d56f3ce12