waf: fix using correct python version for poetry env on configure.
When the default python version is not compatible with the poetry
configuration, and when no specific python version is set on the poetry
environment, poetry complains on each command with the following
message:
    The currently activated Python version 3.x is not supported by the project (xxx).
    Trying to find and use a compatible version.
    Using python3.x (3.x.x)

This message is inserted on every command run with `poetry run` and
breaks our build system.

The best way to solve this issue, is to force using a specific python
version in the poetry environment with `poetry env use 3.x`.

Fortunately, we can easily get the version automatically selected by
poetry and force this version to be used in the poetry environment.

This way, all subsequent commands `poetry run` will not show the message
above and the build system is ok.

Change-Id: Ieebfc054a850c7952606ae27a21901a0019a068a
Priv-Id: 150dcdf502f1addf5e9c8ed70284e6151041a0f5