qps/bitmap: fix safety issue with QPS bitmap enumerator
Squash commit of multiple commits:

---

qps/bitmap: add a bench tool to check enumerator performance

Add a bench tool in order to avoid unexpected performance drop while
cleaning code and fixing stuff in QPS bitmap enumerator.

Change-Id: I7dbe27903a4f45b2d7c211ad884547bbfe511ea9
Priv-Id: 82e79fcc115a065a56f78a760a4c735b597ad84d

---

qps/bitmap: remove supposedly unreachable code

In 'qps_bitmap_enumerator_next_nn()', the '_nn' suffix stands for "non
nullable" and the code handling nullable bitmap is unreachable.

This patch removes the unreachable code from the function and adds an
assert to detect any bad use of the function.

Change-Id: I9284b317ae00d0c364414193d0aa2aec8c7d99ea
Priv-Id: ebdd6876b56cbc30601b72db01be170252c1b5a9

---

qps/bitmap: get rid of abusive code duplication for enumerator builder

The QPS bitmap enumerator has a lot of specialized code that allows it
to make it slightly faster at the cost of an important duplication.

For now I'm not planning to remove this duplication for the next/go_to
functions because the bench seems to show that this specialization bring
a small performance gain (1-2% of CPU time).

However it really doesn't make sense to keep the duplicated enumerator
builder because no one uses it and we expect the performance gain to be
even smaller: an enumerator can be built once only.

Note: in order to improve code organization and limit the possibility of
conflicts, the non-specialized builder is moved in place of the removed
one.

Also remove the macro 'qps_bitmap_for_each_nn()'. Never used either.

Change-Id: Iba26a846649a24023dc81fe1e6f388d58acdd051
Priv-Id: b52c042d63bb8c78f3eb894ef8fde65a22dd9815

---

qps/bitmap: fix code duplication in bitmap enumerator

The specialization of QPS bitmap enumerator duplicated an important part
of the enumerator code. Splitting it in two parts:

1. The "generic" part, that handles both nullable and non-nullable
   bitmaps.

2. The "non-nullable" part, that duplicates the whole non-nullable part
   of the generic part and the global logic of the implementation.

We could have removed the whole duplication by keeping only the generic
part but the bench shows that it could have implied a small but real
performance drop.

Instead of doing this, this patch splits the code in two "symmetrical"
parts: the non-nullable and the nullable parts.

Consequences:

1. The non-nullable part of the code is no more duplicated.

2. The logic of the implementation is still duplicated. This could be
   fixed using .in.h headers at the cost of a decreased maintainability
   (but the removed duplication would also imply a maintainability
   improvement so it might worth it anyway).

3. The performance of the non-nullable implementation stays the same.
   Creation of a nullable implementation, faster than the generic one.
   The performance of the generic implementation is improved because the
   "is_nullable" attribute is only checked once.

Change-Id: I7cbca4052f9830ff72dc64305fe7abdff00579b8
Priv-Id: 64f7cf3795d3f11345805bca8154c0f9becb554b

---

qps/bitmap: remove useless bitfield in QPS bitmap

There is no reason to use a bit field when we don't even expect the
struct to be smaller.

No performance gain noticed though.

Change-Id: I0cd72b4884b8120c43fc074830e0cfc334422e37
Priv-Id: 7d644f2e8fc8c22e8f2a6c5e4d00f3a22478f33f

---

qps/bitmap: add additional safety check for enumerator

Check for a case of concurrent modification that doesn't look buggy in
the current implementation before we actually rework the whole
enumerator safety.

Change-Id: I0f34d245186584a1b0597fa9032a3166ee0c1530
Priv-Id: 3b746fa79781c0148c81642e36c60bb072438327

---

qps-bitmap: fix safety issue with QPS bitmap enumerator

Before this patch, the QPS enumerator was supposed to be always safe.
IOW it was supposed to support modifications in a QPS bitmap currently
being enumerated. That didn't really work and that's why the
issue #83666 was opened: the code only handled modifications in the
bitmap structure but some smaller modifications were unsafe .

This patch gets rid of the previous safety-handling code and makes the
bitmap track all its modifications, not only the ones that change the
structure.

It also adds a "safe" parameter on next and go_to APIs, making the
functions faster in cases where we don't expect the bitmap to change.

Contrarily to the older safety code, we only have a single way to
"refresh" our position in the bitmap. It looks like the QPS bitmap
creators had something more efficient in mind but they left absolutely
no documentation about it.

Note: the code specialization between nullable and non-nullable part is
really annoying, making some parts of this patch twice as big as they
should be. I decided to keep the specialization for now but I'll get rid
of it as soon as I can.

====================================
Bench report (ztst-qps-bitmap-bench)
====================================
Tested on my dev laptop. All tests were run several time to eliminate
possible anomalies. All tests were run with -r5 (five scan repetitions).

We're comparing the benchmark result between the patch implementing the
benchmark (old) and this patch (new). The benchmark tool was build with
P=release.

╔══════════╦════════════════╦═══════════════╗
║ bitmap   ║ enumerator     ║   scan time   ║
╟──────────╫─────────┬──────╫───────┬───────╢
║ nullable ║ generic │ safe ║ old   │ new   ║
╠══════════╬═════════╪══════╬═══════╪═══════╣
║ no       ║ yes     │ yes  ║ 3.30s │ 3.15s ║
║          ║ no      │ yes  ║ 3.18s │ 3.15s ║
║          ║ yes     │ no   ║ N/A¹  │ 2.75s ║
╟──────────╫─────────┼──────╫───────┼───────╢
║ yes      ║ yes     │ yes  ║ 4.99s │ 4.62s ║
║          ║ no      │ yes  ║ N/A²  │ 4.60s ║
║          ║ yes     │ no   ║ N/A¹  │ 4.36s ║
╚══════════╩═════════╧══════╩═══════╧═══════╝
¹ There was no unsafe enumerator implementation before this patch.
² There was no nullable enumerator specialization when the benchmark
  was implemented.

In the benchmark context, the new implementation is always faster, and
the addition of the unsafe mode even bring a significant performance
gain. Of course, the context of the bench can be considered as very
different from a real-life use case but these figures are encouraging
and helps us to consider a performance regression as unlikely.

Change-Id: I3092396e1cabdc369a9b9aacd7cabbba248077d1
Priv-Id: 492079de7342df77d6fed99fde1ad54ea3059167

---