waf/poetry: fix default locale used when forcing lang env variables.
Since the support of poetry in lib-common 29df471cca3, when the
environment `LC_ALL` or `LANG` were not set, we set them to
'en_US.UTF-8' for `click`.

Unfortunately, this implied that the locale 'en_US.UTF-8' was installed,
which is not always the case.

Instead use 'C.UTF-8' which is the POSIX standards-compliant default
locale.

Change-Id: I947a513d23025011d0696085f5218b59e2d9b9aa
Priv-Id: 1c8a21c1a68fc304ff7c3c085eb8476eed4a26c8