waf/poetry: fix setting locale can break configure.
Before, we were setting `LC_ALL` and `LANG` to a default locale for
`click` if they were not set.

1c8a21c1a68fc was supposed to fix the case 'en_US.UTF-8' is not
available and use the more standard 'C.UTF-8'.

Unfortunately, 'C.UTF-8' is not sometimes installed so we have the same
issue as with 'en_US.UTF-8'.

Actually, the recent versions of `click` do not required `LC_ALL` and
`LANG` to be set.

So the solution is to simply do nothing and not setting `LC_ALL` and
`LANG` to a default locale.

Change-Id: I3c55cb3b1e6f5431ad1a485c5e54661ea52de395
Priv-Id: b3332e6bc1180f8364860a4927b33cfd6aa941ba