iopy: fix unexpected exception handling on server callbacks.
Before, when an unexpected error was triggered, the error was reraised
on the main thread without any control where it was actually sent.

This was a big issue, where it can completely stop or block the main
thread in unexpected circumstances.

To fix it, we now have an unexpected exception callback in the server,
similar to `on_connect` callback, to properly handle the exception.
If the callback is not set, or if the callback raises an exception
itself, the exception is logged as a warning.

Change-Id: If0f8a6ebb881f65aee0c3cdae401e76036aa05f5
Priv-Id: 1e681b17ca7c1ba834ae5b07dfdaf4e3803a7dbf