build: version.sh, fix case where version number is 0
In 2e528e8a13b2, version.sh script was fixed to remove leading 0 from
version number. Reason is that if the script output 08 or 09, the
compiler will interpret that as an octal number.
Problem is that the fix break the case where the version number is 0.
In that case, the script would output an empty value which would break
the build.
This patch does check the version number and replace it with 0 if the
value is empty.

Change-Id: I9b5fc8edd6714939d0ea320a7d8dafb5a565f45d
Priv-Id: 612623c18f73e1cf1f49624c5d312e270c63f427