waftools/backend: fix `source` stlib tgen for fpic for waf-2.0.13.
Waf-2.0.13 expects `source` to be set when creating a stlib TaskGen.

Unfortunately this was not longer the case with 0f32982c02fb9aa9 where
we copied the attributes after creating the stlib TaskGen.

To solve this issue, pass the copied attributes directly in the stlib
TaskGen constructor instead of setting them after the stlib TaskGen
being constructed.

Change-Id: Ie9bb4437034a7382de368e5f1602969ceae2ff65
Priv-Id: 82dda8a513295dcfe2e0d52501c4c72ac21be4f6