waftools: add an option for shared libraries to keep the 'lib' prefix
For the packaging of MCG, the option "keep_lib_prefix" was added to
directly create the binary with the lib prefix.

Usually when building a shared library, the prefix 'lib' is used as
prefix in the generated file name. In our waf integration, there is
a function that removes this prefix.

In cases where we have a target that depends on a shared library, the
shared library will be built without the prefix and the target will
fail to be built because it expects a file with the 'lib' prefix.

It is mostly not a problem since when we have a target that depends on
a library built by the same sources we use static library instead.

When the dynamic library was required, we used a symbolic link with the
'lib' prefix to point to the actually built library without the prefix.

This patch adds an option to keep the prefix in the library name.
By default, it doesn't change the behavior but if enabled, the 'lib'
prefix will be kept.

This patch is used in I9f1e79e692faa90266d21aac25cce7afc3bb2f5b.

Change-Id: Icc5ca159b7af36fec60dc97e312aeb660e3b61cc
Priv-Id: 0e8bc87530250d71a94d2b532b2cb817f0204a7f