qps: Change snapshot_syn handling.
The QPS snapshot_syn can be used by external libraries to post jobs that
should be awaited before ending a snapshot.

It is not used internally, and most of time is not used at all, so
having it owned by the qps_t is not really appropriate.

This patch modifies that, by considering that qps_t is not the owner of
the thr_syn_t, and have a function to set it instead of a getter.
This will also allow the users of this library to change the thr_syn_t
used by a qps_t.

Change-Id: I367738313f35d5964eb49e8187957e04741dfb73
Priv-Id: 1ab09237b7a4270fb98f915727d8d146eff42d9c