waftools, add clang_compilation_database extension
clangd is a tool from clang that allows editor to retrieve
informations like documentation, or to do code linting.
In order to work correctly, clangd needs to known the compile command
needed for a file. A simple way to configure clangd is to generate a
compile_commands.json file. This file lists all files of the repo and
for each file gives the command used to build the file.

While waf does have an extension to generate this file, we cannot use
it:
* the extension is available only in recent versions. So older
  installation would not have it.
* we have features like block which would not work correctly with the
  extension from waf.

This patch adds a modified version of the waf extension to the build
system. For the moment, the changes are only to follow intersec coding
rules. In next patch, we will implement block support and discard
generated files.

Change-Id: If1c289c2e0ac575e62763abfd7fde6f9f7528b70
Priv-Id: f093442cfb417b59e2ddad0495e67c86442ac2af