iopy: Add a soft watch activity on client ichannels.
Since mmsx commit ea534a4d6bf, the product server ichannels have a
default connection timeout for public peers (eg. iopy).

So ichannels opened by the python that were inactive for 60 seconds (by
default) were closed by the backend. We thought about this problem but
consedered it was not an issue because iopy automatically reconnects the
ichannels when needed.

But there is a case where it is a problem: when a single RPC call takes
more than 60 seconds to reply (we have this kind of things in slow
tests). In that case, the RPC fails with an ABORT, which can make de
tests fail. This is what happens in #89644.

To fix that, just hardcode a 10 seconds soft watch activity on all
client ichannels. Thanks to that, the ichannel will generate regular
pings, maintaining the connection alive), without ever taking the
decision to disconnect it. IC_EVT_ACT and IC_EVT_NOACT must be properly
ignored in iopy_ic_client_on_event then.

Change-Id: I987950b865b05db22349b0d5ff74c1aae3c27b4f
Priv-Id: 1e2a7fd8ab9dc4345c4e1dff0b067dc7312db2b4