iop: allow catching errors when copying too big IOPs
When copying/duplicating an IOP struct that needs a memory chunk bigger
than 1GiB to be written, the IOP library crashes (panic) with the
infamous error message:

 "you cannot allocate that amount of memory [...]".

In some cases, that case is not so unlikely to happen and we would like
to handle this kind of errors instead of crashing, but with no change in
the IOP library, it would mean to compute the allocation size twice,
which would not be efficient.

This patch adds a new copy flag named IOP_COPY_PROTECT that allows the
duplication/copy function to stop in a clean way when the IOP object is
too big.

Change-Id: Ibe20765c70079eefc2de71ea6614982a9ca0cd3c
Priv-Id: 7851c558d73226712aacbe53510ca7eb9da02610