asdf: move make dependency to the first place
Having the make dependency before the python one helps during the first
setup of the repository.

Indeed depending on the ASDF setup of the user, the python installation
may fail because python requires make to be built. If ASDF isn't
authorized to fallback to the system make (which isn't the default) when
the plugin isn't installed yet the ASDF installation will fail until the
make plugin is manually installed.

Inverting the dependencies fix this issue.

Change-Id: Idd4c0307d78965a0ec4bdffc06a171ab9fc0e175
Priv-Id: 94881bfdc895a884b84c56d814fd4d06030f3af2