core-macros: fix NEXTARG compilation issue with GCC 12.x
The NEXTARG() is designed to be used systematically without thinking to
always be safe when adding new arguments. So it's often use to read
a single argument, leading to "set but not used" compilation errors,
like this one:

    bigdata/ctl.c:83:43: error: parameter 'argc' set but not used [-Werror,-Wunused-but-set-parameter,Semantic Issue]
    static int cmd_cfg_update_schema_impl(int argc, char **argv)
                                              ^

To silent GCC we add an intermediate `(void)argc` in the macro that will
be removed at compile time and which is the standard way to disable the
"set but not used" warnings.

Change-Id: If3fd6f90f5337a223b1be8cd0a875505737a6914
Priv-Id: f241e1f8eb754aafd277237ced87036aaf09e893