waf: Ensure binaries / shared-libraries are re-linked after configure.
When changing the build profile (for example changing from debug to
asan), and then switching back, the binaries were not re-linked because
nothing changed from the previous run (in the same profile).

So we could easily compile in a profile (let's say asan) but actually
end up with binaries of the previous build on another profile (for
example debug), without noticing. This can be a waste of time, and even
prevent from detecting memory bugs (in that particular case).

Fix that by adding the configure time in the "hcode" of the link tasks,
which will force a re-link after each configure. The problem with that
is that the link will be remade even if the profile was actually not
change, and that there is nothing to do. But I could not find something
smarter, and this should not be a big problem.

Change-Id: I429f7fd580da6227bff5ff0b030f2eeb1232931e
Priv-Id: 05fb825c70037aada4267f43f5b97722622ec757