waf/asdf: always rerun waf with ASDF's poetry
When using ASDF we need to handle correctly a change of python version
between two `waf configure` (which is the whole point of using ASDF).

But before this patch, if you were launching `waf configure` from an
already activated venv, the venv would enforce its python version which
is not what we want when using ASDF. We want to use the version
committed in the .tool-versions and then switch the current venv to
this version.

Change-Id: Iaa2337f6fb8e9a29c842111d661e72795fabe1e6
Priv-Id: 146c02b0b1fcfc746cec838de261d69ac65a9f30