qps/hat: undo a change that uncovered a bug yet to investigate
Undo the patch 18b173de6284f17b7d9f2b79a4b5b967ac565666 because it was
identified as the cause of some QHAT corruptions (issue #94699). We
believe that the root cause of the corruption is different and that the
patch only uncovered the bug, but undoing it is the best option we have
until we identify that root cause.

Change-Id: Ifc491d525b297c850e5c5949d8f95e5e9168fbe8
Priv-Id: 8f545ea59cad36162a353623c1a99a64d7672350