iop/zchk: fix missing shlibs deps in tests
In order to succeed, the test "zchk iop" needed three shared libs to be
built but these dependencies were not declared in the wscript. As a
consequence, if we only built "zchk" which "waf --target=zchk", the IOP
tests failed.

Also remove unneeded include from "zchk-iop.c": we did not need
"tstiop2.iop" for compilation: we only needed the shared library
"zchk-tstiop2-plugin" for tests success.

Change-Id: Ic4d68e0eb767b0406b613a962e0c758268db423c
Priv-Id: fb88d3cfd7858bac92446f9f353b2e2c85fa792c